Добавлены тесты (по неизвестным причинам sqlite не работает правильно на эмуляторе)

+++ adfmp1h21bird/app/src/main/java/com/example/adfmp1h21_bird/ui/note_page/NoteFragment.kt
@@ -96,7 +96,7 @@ class NoteFragment : Fragment() {
     private fun getData(NoteID:String): MyNote {
 
         val temp = MyNote(
-                NoteID.toLong(),
+                NoteID.toInt(),
                 "Неро $NoteID",
                 R.drawable.test.toString(),
                 "Какой-то geotag",
